First set of unit tests

+++ src/main/java/com/jaehyune/stamp_app/service/impl/CommentServiceImpl.java
@@ -60,6 +60,9 @@ public class CommentServiceImpl implements CommentService {
     @Transactional
     @Override
     public CommentReadDTO findById(Integer id) {
+        if (id == null || id < 0) {
+            throw new RuntimeException("Invalid ID: " + id);
+        }
         Optional<Comment> comment = commentRepository.findById(id);
 
         if (comment.isPresent()) {
@@ -79,6 +82,9 @@ public class CommentServiceImpl implements CommentService {
     @Transactional
     @Override
     public List<CommentReadDTO> findAllCommentForStamp(Integer stamp_id) {
+        if (stamp_id == null || stamp_id < 0) {
+            throw new RuntimeException("Invalid ID: " + stamp_id);
+        }
         Optional<Stamp> temp = stampRepository.findById(stamp_id);
         Stamp stamp;
         if (temp.isPresent()) {
@@ -92,9 +98,13 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     public void delete(Integer id) {
+        if (id == null || id < 0) {
+            throw new RuntimeException("Invalid ID: " + id);
+        }
         commentRepository.deleteById(id);
     }
 
+    // TODO: Lombok
     public Comment toEntity(CommentCreationDTO dto) {
         Integer userId = dto.getUserId();
         Optional<User> tempUser = userRepository.findById(userId);
